requirements.txt
backports.shutil-get-terminal-size==1.0.0
click==6.6
decorator==4.0.10
Flask==0.11.1
flask-peewee==0.6.7
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.0
pandas==0.18.1
pathlib2==2.1.0
peewee==2.8.1
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
python-dateutil==2.5.3
pytz==2016.4
requests==2.10.0
sh==1.11
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
Werkzeug==0.11.10
wtf-peewee==0.2.6
WTForms==2.1
